requirements.txt
python-dotenv==0.20.0
certifi==2024.6.2
python-dotenv==0.20.0

transformers==4.40.0
SentencePiece==0.2.0
torch==2.3.1
datasets==2.20.0

# rich
git+https://github.com/huggingface/parler-tts.git
descript-audio-codec==1.0.0

soundfile==0.12.1
librosa==0.10.2.post1
numpy==1.26.4